Refactor Structs To Subdirectory
