+++ src/structs.rs
@@ -1,0 +1,3 @@
+pub mod hitbox;
+pub mod roomcache;
+pub mod taskdata;
