+++ src/structs/roomcache.rs
